+++ src/file-funcs.js
@@ -1,7 +1,7 @@
 module.exports = {
     save: function (fileName, contents) {
         try { 
-            require("fs").writeFileSync(fileName, contents); 
+            require ("fs").writeFileSync (fileName, contents); 
             return true;
         } catch (error) { 
             return false;