+++ changelog.md
@@ -14,4 +14,4 @@
     - migrated server side code to Node.js  
     - content boundary size relative to window width  
     - zero dependencies file picker  
-    
+    - input/output menu